Fix Casts

+++ tsMuxer/bufferedFileReader.h
@@ -2,7 +2,6 @@
 #define BUFFERED_FILE_READER_H_
 
 #include <fs/file.h>
-#include <fs/systemlog.h>
 #include <types/types.h>
 
 #include "BufferedReader.h"
@@ -39,13 +38,13 @@ struct FileReaderData final : ReaderData
 
     ~FileReaderData() override = default;
 
-    uint32_t readBlock(uint8_t* buffer, const int max_size) override { return m_file.read(buffer, max_size); }
+    uint32_t readBlock(uint8_t* buffer, const uint32_t max_size) override { return m_file.read(buffer, max_size); }
 
     bool openStream() override;
     bool closeStream() override { return m_file.close(); }
     bool incSeek(const int64_t offset) override
     {
-        return m_file.seek(offset, File::SeekMethod::smCurrent) != static_cast<uint64_t>(-1);
+        return m_file.seek(offset, File::SeekMethod::smCurrent) != -1;
     }
 
     File m_file;
@@ -57,9 +56,9 @@ class BufferedFileReader final : public BufferedReader
    public:
     BufferedFileReader(uint32_t blockSize, uint32_t allocSize = 0, uint32_t prereadThreshold = 0);
 
-    bool openStream(uint32_t readerID, const char* streamName, int pid = 0,
+    bool openStream(int readerID, const char* streamName, int pid = 0,
                     const CodecInfo* codecInfo = nullptr) override;
-    bool gotoByte(uint32_t readerID, uint64_t seekDist) override;
+    bool gotoByte(int readerID, int64_t seekDist) override;
 
    protected:
     ReaderData* intCreateReader() override { return new FileReaderData(m_blockSize, m_allocSize); }
